Touched up maze generation and score
Maze generator never turns in the same direction twice to reduce looping
and to prevent near dead ends (like normal dead ends but with a loop at
the end so there isn't a stop but there is still only one way out).
Also, the score for eating ghosts scales properly now so that the player
gets twice as many points for every ghost eaten off of a single big dot
(ie 200 if one ghost is consumed, 400 for the next, 800 for the next,
etc. and it resets after the power wears off).