Renamed PacEntity class to Player
My first choice would be 'PacMan', but I already have a class named that
because it is the name of the game. I could still use it because it is
in different package, but that would have been confusing. I named it
PacEntity originally, but that was also confusing, so I changed it to
player.