Modified Maze Generation and Power Dots
I modified the "fix dots" method in maze.java to make it more general
and capable of fixing almost all dots. Also, I declared a constant in
Player.java to use for the power gained from big dots so that it can
once again be changed by modifying one line of code.